docs/examples/solution/requirements.txt
geopy==2.2.0
matplotlib==3.3.4
numpy==1.19.5
paho-mqtt==1.6.1
shapely==1.8.2